Adjusted File Placement and Added Player Data Objects
+ Placed the scripts into the develop folder
+ Added the scriptable objects for the player data
