+++ Project B/Assets/Plugins/Sirenix/Demos/Custom Attribute Processors.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f597f19f656ba56eae4f6a3a7cc528f4
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
