+++ Project B/Assets/Plugins/Sirenix/Demos/Editor Windows.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ed09910c0094cb27be8f3ca264680da3
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
