+++ Project B/Assets/Plugins/Sirenix/Demos/Sample - RPG Editor.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: cc355dd4cf1e6173beaeb22c2858cbe1
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
