+++ Project B/Assets/_Develop/Player Features/Scriptable Assets.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 818bdd4ee2ffb3e4fa6bb51ca87b237a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
