+++ Project B/Assets/_Develop/Player Features/Scriptable Assets/Arrow Input.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 98f8546cdce9cca4b8a2a6cf274d22ed
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
