+++ Project B/Assets/_Develop/Player Features/Scriptable Assets/WASD Input.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aae629f5f36dd064e8cb137a3e1c20db
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
