+++ Project B/Assets/_Develop/Player Features/Scripts/Data/PlayerInput.cs
@@ -1,0 +1,8 @@
+﻿using UnityEngine;
+
+[CreateAssetMenu]
+public class PlayerInput : ScriptableObject
+{
+    [SerializeField] public KeyCode Up, Down, Left, Right;
+    [SerializeField] public KeyCode Place;
+}
