+++ Project B/Assets/_Develop/Player Features/Scripts/Data/PlayerInput.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6ed78a6a34ab34c4ca410944b4b38ae0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
